ADD: Display Lightning details in Invoice View
