+++ class/wallets/types.ts
@@ -81,13 +81,19 @@ export type TransactionOutput = {
 
 export type LightningTransaction = {
   memo?: string;
-  type?: 'user_invoice' | 'payment_request' | 'bitcoind_tx' | 'paid_invoice';
+  type?: "user_invoice" | "payment_request" | "bitcoind_tx" | "paid_invoice";
   payment_hash?: string | { data: string };
-  category?: 'receive';
+  category?: "receive";
   timestamp?: number;
   expire_time?: number;
   ispaid?: boolean;
   walletID?: string;
+  value?: number;
+  amt?: number;
+  fee?: number;
+  payment_preimage?: string;
+  payment_request?: string;
+  description?: string;
 };
 
 export type Transaction = {
